requirements.txt
langgraph>=0.2.60
langgraph-checkpoint-sqlite>=2.0.0
langchain-openai>=0.2.14
langchain-anthropic>=0.2.4
tavily-python>=0.5.0
PyMuPDF>=1.24.14
httpx>=0.28.1
python-dotenv>=1.0.1
langsmith>=0.1.147
websockets>=13.0

# Web framework
fastapi>=0.104.0
uvicorn>=0.24.0
jinja2>=3.1.2
python-multipart>=0.0.6
